--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice02/s21K0647/TestBall.java
@@ -1,21 +1,0 @@
-package ru.mirea.practice02.s21K0647;
-
-
-import java.util.Scanner;
-
-public class TestBall {
-    public static void main(String[] args) {
-
-        Ball b1 = new Ball(0, 0);
-        System.out.println(b1.toString());
-        Scanner source = new Scanner(System.in);
-        System.out.print("Ball x and y: ");
-        b1.setX(source.nextDouble());
-        b1.setY(source.nextDouble());
-        System.out.println(b1.toString());
-        System.out.print("Ball move x and y : ");
-        b1.move(source.nextDouble(), source.nextDouble());
-        System.out.print("Ball move x and y : ");
-        System.out.print(b1.toString());
-    }
-}
