--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice03/s21K0647/Circle.java
@@ -1,25 +1,0 @@
-package ru.mirea.practice03.s21K0647;
-
-public class Circle extends Point {
-
-    Point T;
-    double r;
-
-    public Circle(){
-        r = 0;
-    }
-
-    public Circle(double r){
-        this.r = r;
-    }
-
-
-    public double getX() {return x; }
-
-    public void setX(double x) {this.x = x; }
-
-    public double getY() {return y; }
-
-    public void setY(double y) {this.y = y;}
-
-}
