--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice03/s21K0647/Tester.java
@@ -1,22 +1,0 @@
-package ru.mirea.practice03.s21K0647;
-
-import java.util.Scanner;
-
-public class Tester {
-    public static void main(String[] args) {
-        Scanner source = new Scanner(System.in);
-        int n;
-        n = source.nextInt();
-        Circle []c = new Circle[n];
-        for (int i = 0; i < n; i++) {
-            Circle a = new Circle();
-            a.x =  source.nextDouble();
-            a.y =  source.nextDouble();
-            a.r =  source.nextDouble();
-            c[i] = a;
-        }
-        for (Circle p: c) {
-            System.out.println("x: " + p.x + "\ty: " + p.y + "\tRadiuse: " + p.r);
-        }
-    }
-}
