--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice04/s21K0647/Shop.java
@@ -1,30 +1,0 @@
-package ru.mirea.practice04.s21K0647;
-
-import java.util.ArrayList;
-public class Shop {
-    ArrayList<Computer> C = new ArrayList<Computer>();
-
-    public void addComputer(Computer a) {
-        C.add(a);
-    }
-
-
-    public void deleteComputer(String model) {
-        for (Computer x : C) {
-            if(x.model.equals(model)){
-                C.remove(x);
-            }
-        }
-    }
-    public String searchComputer(String model) {
-        String n = "Not found";
-        for (Computer x : C) {
-            if (x.model.equals(model)) {
-                n =  "Model: " + x.model + "\tProcessor: " + x.processor + "\tMemory: " + x.memory + "\tBattery: " + x.battery;
-            }
-        }
-        return n;
-    }
-}
-
-
