--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice05/s21K0647/Dog_kennel.java
@@ -1,19 +1,0 @@
-package ru.mirea.practice05.s21K0647;
-
-import java.util.Scanner;
-import java.util.ArrayList;
-
-public class Dog_kennel {
-    public static void main(String[] args) {
-        Scanner source = new Scanner(System.in);
-        Dog a1 = new Dog("Jerry", 10);
-        ArrayList<Dog> kennel = new ArrayList<Dog>();
-        kennel.add(a1);
-        Dog a2 = new Dog("Pat", 4);
-        kennel.add(a2);
-        System.out.println(kennel.get(1).ManAlt());
-        for(Dog x: kennel){
-            System.out.println(x.ToString());
-        }
-    }
-}
