--- students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice07/s21K0647/Book.java
@@ -1,25 +1,0 @@
-package ru.mirea.practice07.s21K0647;
-
-public class Book {
-    String author;
-    String title;
-    int year;
-
-    public Book(String author, String title, int year) {
-        this.author = author;
-        this.title = title;
-        this.year = year;
-    }
-
-    public String getAuthor(){ return author;}
-
-    public void setAuthor(String author){ this.author = author;}
-
-    public String getTitle(){ return title;}
-
-    public void setTitle(String title){ this.title = title;}
-
-    public int getYear(){ return year;}
-
-    public void setYear(int year){ this.year = year;}
-}
