+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice1/s21K0647/TestAuthor.java
@@ -1,0 +1,12 @@
+package ru.mirea.practice1.s21K0647;
+
+
+import java.util.Scanner;
+
+public class TestAuthor {
+    public static void main(String[] args) {
+        Scanner source = new Scanner(System.in);
+        Author k1 = new Author(source.nextLine(), source.nextLine(), source.next().charAt(0));
+        System.out.print(k1.toString());
+    }
+}
