+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice2/s21K0647/Ball.java
@@ -1,0 +1,34 @@
+package ru.mirea.practice2.s21K0647;
+
+public class Ball {
+
+    double x = 0.0;
+    double y = 0.0;
+
+    public Ball(){};
+    public Ball(double x, double y){
+        this.x = x;
+        this.y = y;
+    }
+
+    public double getX(){ return x;}
+
+    public void setX(double x){ this.x = x;}
+
+    public double getY(){ return y;}
+
+    public void setY(double y){ this.y = y;}
+
+    public void setXY(double x, double y){ this.x = x; this.y = y;}
+
+    public void move(double xDisp, double yDisp){
+        this.x += xDisp;
+        this.y += yDisp;
+    }
+
+    @Override
+    public String toString() {
+        return "Ball x, y: " + x + " " + y;
+    }
+
+}
