+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice3/s21K0647/Point.java
@@ -1,0 +1,23 @@
+package ru.mirea.practice3.s21K0647;
+
+public class Point {
+    double x;
+    double y;
+
+    public Point(){
+        x = 0;
+        y = 0;
+    }
+
+    public Point(double x, double y){
+        this.x = x;
+        this.y = y;
+    }
+    public double getX(){return x;}
+
+    public void setX(double x){ this.x = x;}
+
+    public double getY(){return y;}
+
+    public void setY(double y){ this.y = y;}
+}
