+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice4/s21K0647/Computer.java
@@ -1,0 +1,32 @@
+package ru.mirea.practice4.s21K0647;
+
+public class Computer {
+    String model;
+    String processor;
+    String memory;
+    String battery;
+    public Computer(String model, String processor, String memory, String battery){
+        this.model = model;
+        this.processor = processor;
+        this.memory = memory;
+        this.battery = battery;
+    }
+
+    public String getModel(){ return model;}
+
+    public void setModel(String model){ this.model = model;}
+
+    public String getProcessor(){ return processor;}
+
+    public void setProcessor(String processor){ this.processor = processor;}
+
+    public String getMemory(){ return memory;}
+
+    public void setMemory(String memory){ this.memory = memory;}
+
+    public String getBattery(){ return battery;}
+
+    public void setBattery(String battery){ this.battery = battery;}
+
+
+}
