+++ students/21K0647/21K0647-p02/src/main/java/ru/mirea/practice9/s21K0647/Karta.java
@@ -1,0 +1,19 @@
+package ru.mirea.practice9.s21K0647;
+
+public class Karta {
+    String weight;
+    String mast;
+    public Karta(){
+        weight = "";
+        mast = "";
+    }
+    public Karta(String weight, String mast){
+        this.weight = weight;
+        this.mast = mast;
+    }
+    public void setWeight(String weight){ this.weight = weight;}
+    public void setMast(String mast){ this.mast = mast;}
+
+
+
+}
